+++ x/incentive/types/expected_keepers.go
@@ -3,8 +3,9 @@ package types
 import (
 	"context"
 
-	epochingtypes "github.com/babylonlabs-io/babylon/v3/x/epoching/types"
 	sdk "github.com/cosmos/cosmos-sdk/types"
+
+	epochingtypes "github.com/babylonlabs-io/babylon/v3/x/epoching/types"
 )
 
 type AccountKeeper interface {